+++ packages/mui-base/src/BadgeUnstyled/index.d.ts
@@ -4,5 +4,8 @@ export * from './BadgeUnstyled';
 export { default as useBadge } from './useBadge';
 export * from './useBadge';
 
+export { default as BadgeUnstyledProps } from './BadgeUnstyledProps';
+export * from './BadgeUnstyledProps';
+
 export { default as badgeUnstyledClasses } from './badgeUnstyledClasses';
 export * from './badgeUnstyledClasses';
